FRU: pldmresponder: fix width of PLDM entity type
One of the APIs had the entity type wrongly set at a byte. Entity type
is two bytes.

+++ libpldmresponder/fru_parser.hpp
@@ -22,7 +22,7 @@ using Interface = std::string;
 using Interfaces = std::set<Interface>;
 using Property = std::string;
 using PropertyType = std::string;
-using EntityType = uint8_t;
+using EntityType = uint16_t;
 
 } // namespace dbus
